[refactor] remove magic number and duplication for Read range Error change long type to unsigned

+++ SSDSimulator/SSDSimulator/ssdCmdRead.cpp
@@ -2,7 +2,7 @@
 #include <fstream>
 #include <sstream>
 
-void SsdReadCmd::CheckAddressRange(long newAddress)
+void SsdReadCmd::CheckAddressRange(uint32_t newAddress)
 {
     if (newAddress < 0 || newAddress > 99) {
         std::ofstream outFile("ssd_output.txt");
